# gateway/apidocs/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AccelByte Challenge Service API",
    "description": "Challenge service for managing user challenge progress and reward claims",
    "version": "1.0"
  },
...
  "basePath": "/challenge",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/challenges": {
      "get": {
        "summary": "Get user challenges",
        "description": "Retrieve all challenges for the authenticated user with current progress",
        "operationId": "Service_GetUserChallenges",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serviceGetChallengesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "activeOnly",
            "description": "M3 Phase 4: Filter to show only active goals (default: false shows all goals)",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/v1/challenges/initialize": {
      "post": {
        "summary": "Initialize player goals",
        "description": "Assign default goals to new players or sync existing players with config changes",
        "operationId": "Service_InitializePlayer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serviceInitializeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/serviceInitializeRequest"
            }
          }
        ],
        "tags": [
          "Challenges"
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/v1/users/me/data": {
      "delete": {
        "summary": "Delete user data (GDPR)",
        "description": "Delete all goal progress data for the authenticated user. Used for GDPR right-to-erasure compliance.",
        "operationId": "GDPR_DeleteUserData",
        "responses": {
          "200": {
            "description": "User data deleted successfully.",
            "schema": {
              "$ref": "#/definitions/serviceGDPRDeletionResponse"
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid bearer token.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          },
          "500": {
            "description": "Internal server error.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "GDPR"
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "serviceAssignedGoal": {
      "type": "object",
      "properties": {
        "challengeId": {
          "type": "string"
        },
        "goalId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "assignedAt": {
          "type": "string"
        },
        "expiresAt": {
          "type": "string"
        },
        "progress": {
          "type": "integer",
          "format": "int32"
        },
        "target": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "type": "string"
        },
        "requirement": {
          "$ref": "#/definitions/serviceRequirement"
        },
        "reward": {
          "$ref": "#/definitions/serviceReward"
        }
      }
    },
    "serviceChallenge": {
      "type": "object",
      "properties": {
        "challengeId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "goals": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/serviceGoal"
          }
        }
      },
      "title": "Domain Models"
    },
...
    "serviceGDPRDeletionResponse": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "description": "The user ID whose data was deleted"
        },
        "rowsDeleted": {
          "type": "integer",
          "format": "int64",
          "description": "Number of goal progress rows deleted"
        }
      }
    },
    "serviceGetChallengesResponse": {
      "type": "object",
      "properties": {
        "challenges": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/serviceChallenge"
          }
        }
      }
    },
...
    "serviceGoal": {
      "type": "object",
      "properties": {
        "goalId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "requirement": {
          "$ref": "#/definitions/serviceRequirement"
        },
        "reward": {
          "$ref": "#/definitions/serviceReward"
        },
        "prerequisites": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "progress": {
          "type": "integer",
          "format": "int32"
        },
        "status": {
          "type": "string"
        },
        "locked": {
          "type": "boolean"
        },
        "completedAt": {
          "type": "string"
        },
        "claimedAt": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "expiresAt": {
          "type": "string"
        },
        "expiresInSeconds": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "serviceInitializeRequest": {
      "type": "object",
      "title": "User ID and namespace extracted from JWT (not from request body)"
    },
    "serviceInitializeResponse": {
      "type": "object",
      "properties": {
        "assignedGoals": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/serviceAssignedGoal"
          }
        },
        "newAssignments": {
          "type": "integer",
          "format": "int32"
        },
        "totalActive": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "serviceRequirement": {
      "type": "object",
      "properties": {
        "statCode": {
          "type": "string"
        },
        "operator": {
          "type": "string"
        },
        "targetValue": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "serviceReward": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "rewardId": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}